Refactor tests for comitato Azure retirements
- Remove obsolete test_raw_source_integrity.py and test_service_health_v1.py files.
- Introduce test_selected_set.py to validate selected set paths and artifact integrity.
- Update test_empty_publication.py and test_staged_bytes.py to use new report structures.
- Add new tests for advisor and service health reports in test_advisor.py and test_service_health.py.
- Implement catalog tests in test_catalog.py to ensure proper dependency and publication order.
