+++ src/comitato/comitato_azure_retirements_v2/application/__init__.py
@@ -1,6 +1,4 @@
-"""Application orchestration and dependency planning."""
+"""Application orchestration."""
 from ..domain.evidence import AdvisorEnrichments, ServiceHealthSupplementalEvidence
-from .advisor import normalize_advisor
-from .service_health import normalize_service_health
 
-__all__ = ["AdvisorEnrichments", "ServiceHealthSupplementalEvidence", "normalize_advisor", "normalize_service_health"]
+__all__ = ["AdvisorEnrichments", "ServiceHealthSupplementalEvidence"]
