+++ src/comitato/comitato_azure_retirements_v2/reports/__init__.py
@@ -1,0 +1,1 @@
+"""Report definitions and selector policy for Azure retirements v2."""
